% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s1ocn_the.R
\name{get_attributes_list}
\alias{get_attributes_list}
\title{Retrieve and Cache the List of Acceptable Attribute Names for Sentinel-1}
\description{
This function retrieves a list of acceptable attribute names and their types for the Copernicus Sentinel-1 Mission.
The data is retrieved from the Copernicus Data Space Catalogue via an OData API. To improve performance, the results
are cached after the first retrieval. Subsequent calls will return the cached list unless the cache is reset.
}
\value{
A list of acceptable attribute names and their types for the Sentinel-1 mission, retrieved from the
Copernicus Data Space Catalogue. If the data has already been retrieved during the session, the function returns
the cached version to avoid unnecessary API calls.
}
\details{
The function operates as follows:
\enumerate{
\item \strong{Cache Check:} The function first checks if the list of attributes is stored in the cache (\code{s1ocn_the$cache$attibutes_list}).
If the cache is "zapped" (a state indicating it is not set or has been reset), the function proceeds to retrieve the data.
\item \strong{Data Retrieval:} If the cache is empty, the function opens a connection to the OData API at the URL
\verb{https://catalogue.dataspace.copernicus.eu/odata/v1/Attributes(SENTINEL-1)}.
\item \strong{Data Parsing:} The function reads the JSON data from the API and parses it into a list format.
\item \strong{Caching:} After retrieving and parsing the data, the function stores the result in the cache to improve the efficiency of
subsequent calls.
\item \strong{Return:} The function returns the list of attributes, either from the cache or freshly retrieved.
}

The attributes list includes the names of acceptable attributes for querying the Sentinel-1 dataset, along with their respective types.
}
\keyword{internal}
