% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1ocn_list_files.R
\name{s1ocn_build_attribute_query}
\alias{s1ocn_build_attribute_query}
\title{Build an OData Query for Sentinel-1 Attributes}
\usage{
s1ocn_build_attribute_query(
  attribute_name,
  attribute_value,
  value_operator = "eq"
)
}
\arguments{
\item{attribute_name}{A character string representing the name of the Sentinel-1 attribute to filter by. The attribute name must be one of the valid attributes returned by \code{s1ocn_the$get_attributes_list()}.}

\item{attribute_value}{The value of the attribute to filter by. The type of the value must correspond to the attribute's value type (e.g., string or numeric). If the value is a string, it will be automatically encoded for the query.}

\item{value_operator}{A character string representing the comparison operator to use. Accepted values are:
\itemize{
\item \code{"eq"} (equal)
\item \code{"le"} (less than or equal)
\item \code{"ge"} (greater than or equal)
\item \code{"lt"} (less than)
\item \code{"gt"} (greater than)
Defaults to \code{"eq"}.
}}
}
\value{
A character string representing the OData query formatted for Sentinel-1 attributes, with the specified attribute name, value, and operator.
}
\description{
This function constructs a valid OData query string to filter Sentinel-1 attributes based on a specified attribute name, value, and comparison operator.
It validates the attribute name and constructs the query according to the attribute's type (string or numeric).
}
\details{
The function works as follows:
\enumerate{
\item \strong{Operator Validation:} The provided \code{value_operator} is validated to ensure it is one of the accepted values.
\item \strong{Attribute Validation:} The \code{attribute_name} is validated against the list of acceptable attributes retrieved from \code{s1ocn_the$get_attributes_list()}. If the attribute is invalid, an error is thrown.
\item \strong{Value Type Handling:} The function determines the attribute's value type from the retrieved attributes list. If the value is a string, it is URL-encoded.
\item \strong{Query Construction:} The function constructs an OData query string in the format required for querying Sentinel-1 attributes.
}
}
\examples{
# Build a query for the 'swathIdentifier' attribute with the value 'IW'
s1ocn_build_attribute_query("swathIdentifier", "IW")

# Build a query for the 'relativeOrbitNumber' attribute with the value 125 and operator 'gt'
s1ocn_build_attribute_query("relativeOrbitNumber", 125, "gt")

}
