% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1ocn_list_files.R
\name{s1ocn_list_files}
\alias{s1ocn_list_files}
\title{Retrieve Sentinel-1 Files Based on Filter Criteria}
\usage{
s1ocn_list_files(
  max_results = 20,
  search_polygon = rlang::zap(),
  datetime_start = rlang::zap(),
  datetime_end = rlang::zap(),
  attributes_search = list()
)
}
\arguments{
\item{max_results}{An integer specifying the maximum number of results to return. Defaults to 20. If greater than 1000, pagination will be handled automatically.}

\item{search_polygon}{An optional search polygon in \code{sf}, matrix, data frame, or WKT string format. The polygon defines the geographical area to search. If not provided, no spatial filter is applied.}

\item{datetime_start}{An optional date-time string specifying the start of the date range for the search. Accepts various date-time formats (e.g., "YYYY-MM-DD"). If not provided, defaults to "1900-01-01T00:00:00.000Z".}

\item{datetime_end}{An optional date-time string specifying the end of the date range for the search. Defaults to the current UTC date and time if not provided.}

\item{attributes_search}{A list of key-value pairs representing attributes and their corresponding values for filtering the search. Attribute names must be valid according to \code{s1ocn_the$get_attributes_list()}. The \code{productType} attribute is handled separately and should not be included in this list.}
}
\value{
A data frame of Sentinel-1 files matching the search criteria. If no files are found, the function returns \code{rlang::zap()}.
}
\description{
This function retrieves a list of Sentinel-1 files based on various search filters, such as attributes, date ranges, and geographical areas.
It supports paginated results when the maximum number of results exceeds 1000. The function queries the Copernicus Data Space Catalogue using OData.
}
\details{
The function operates as follows:
\enumerate{
\item \strong{Query Construction:} It constructs an OData query based on the provided filter criteria, including attributes, date ranges, and optional geographical areas.
\item \strong{Pagination:} If \code{max_results} exceeds 1000, the function automatically handles pagination by iterating through multiple result pages.
\item \strong{Data Retrieval:} The function sends a query to the OData API of the Copernicus Data Space Catalogue, reads the JSON response, and processes the result.
\item \strong{Empty Result Handling:} If no files are found, the function returns \code{rlang::zap()} to signify an empty result.
}
}
\examples{
\dontrun{
# Retrieve Sentinel-1 files with specified attributes
s1ocn_list_files(attributes_search =
list(swathIdentifier = "IW", orbitDirection = "DESCENDING", relativeOrbitNumber = 125))

# Retrieve files with attributes and a date range filter
s1ocn_list_files(attributes_search =
list(swathIdentifier = "IW", orbitDirection = "DESCENDING", relativeOrbitNumber = 125),
datetime_start = "2021-01-01")

# Retrieve files with attributes, date range, and geographical search polygon
search_polygon <- matrix(c(-11, 41, -11, 44, 6, 44, 6, 41, -11, 41), ncol = 2, byrow = TRUE)
s1ocn_list_files(attributes_search =
list(swathIdentifier = "IW", orbitDirection = "DESCENDING", relativeOrbitNumber = 125),
datetime_start = "2021-01-01",
search_polygon = search_polygon)
}

}
