% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1ocn_extract_wind_data.R
\name{s1ocn_wind_data_list_to_tables}
\alias{s1ocn_wind_data_list_to_tables}
\title{Convert a list of wind data into tables}
\usage{
s1ocn_wind_data_list_to_tables(wind_data_list, workers = 1)
}
\arguments{
\item{wind_data_list}{A list of wind data objects from
\code{\link[=s1ocn_extrat_wind_data]{s1ocn_extrat_wind_data()}}.}

\item{workers}{Number of parallel workers to use.}
}
\value{
A list of \code{S1OCN_wind_data_table} objects.
}
\description{
Transform each element of a wind data list into a
\code{\link[=s1ocn_new_S1OCN_wind_data_table]{S1OCN_wind_data_table}}.
}
\details{
The conversion is executed in parallel using \code{furrr::future_map()}.
}
\examples{
\dontrun{
wind <- list(
  a = s1ocn_extrat_wind_data("path/to/a.zip"),
  b = s1ocn_extrat_wind_data("path/to/b.zip")
)
tbls <- s1ocn_wind_data_list_to_tables(wind, workers = 2)
}

}
